core: refactor by splitting impl for each model (#371)
A quick bookmark to browse core api impl before this refactor is [1] but
the entire history can be browsed in the repo/on github.

[1]: https://github.com/kubeflow/model-registry/blob/969056f07f08925ec7b23f67d84cee6a11b58078/pkg/core/core.go
